+++ src/app/dataset-view/additional-components/overview-component/overview.component.ts
@@ -235,12 +235,16 @@ export class OverviewComponent extends BaseComponent implements OnInit {
     }
 
     public get showAddDataButton(): boolean {
-        return (
-            Boolean(this.currentState?.data.length) &&
-            this.isUserLogged &&
-            !this.currentState?.overview.metadata.currentPollingSource &&
-            this.datasetBasics.kind === DatasetKind.Root
-        );
+        if (Boolean(this.currentState?.data.length) && this.isUserLogged) {
+            return (
+                (!this.currentState?.overview.metadata.currentPollingSource &&
+                    this.datasetBasics.kind === DatasetKind.Root) ||
+                (!this.currentState?.overview.metadata.currentTransform &&
+                    this.datasetBasics.kind === DatasetKind.Derivative)
+            );
+        } else {
+            return false;
+        }
     }
 
     public openInformationModal() {
